Crate tic_tac_toe

Structs

Player 🔒

Enums

Piece 🔒

Functions

draw_board 🔒

The draw_board function draws the board array in the terminal. The integers within the array are the valid values players can input.

initialize_board 🔒

The initialize_board function will display an empty board in the terminal.

main 🔒

The main function first initializes the board. It then assigns each player an X or an O. It then draws the board in the terminal. It loops players making moves and redraws the board after each move. After each redraw, it checks if a player has won or if there is a tie and prints a message depending on which.

three_in_a_row 🔒

The three_in_a_row function will determine if three X’s or O’s occur in a row.

tie 🔒

The tie function determines whether or not there is a tie game.

win 🔒

The win function determines if a player has won by using the three_in_a_row function to test all possible orientations in which a player can win.