Module tic_tac_toe::board

With this file we process all the data in the game. We manage the whole game with the functions inside.

Structs

Board

We manage the whole game with the ‘board’ structure. Writing the letters X or O on the grids is the structure where we process which player’s turn it is, or the determination of the winner at the end of each game.

Functions

board_notation

The notation we initially saw while the game was running. It only works once the game starts.

create_board_struct

It returns a ‘board’ structure that we set earlier. Details are below.

restart_board_notation 🔒

If players want to play the game again, the notation they see.

restart_decision 🔒

With this code, we return the first value entered from the keyboard as char.

run_game

The function that runs the game by collecting most functions in its own loop.

score_notation 🔒

Function that displays static variables in the constants.rs file as a scoreboard at the end of each game.

variable_clearing 🔒

If the game will be restarted after each game is over; all variables need to be cleared. In this way, when the game is started, the data of the winning player is set to ‘None’ to prevent possible bugs. It takes a reference &mut board as a parameter.