Function tic_tac_toe::board::run_game

pub fn run_game(board: Board)

The function that runs the game by collecting most functions in its own loop.

When the game starts, we first ask the user to enter a value in the code. Then, we pass this value through various evaluations. Because; text or number may not be entered in the value range we want. Thus, unwanted bugs in the game are avoided.

When the process is finished, the selected section is printed with the for loop below. This code printing current situation.

println!("-------------");
for row in &board.grid {
    ..
    println!("-------------");
}

At the end of each operation, it is checked whether all columns are sorted. If it is what we want, board.winner is set to the value in that column. Then we check this with match.

match board.winner {
    Some(Player::X) => { .. }
    Some(Player::O) => { .. }
    None => { .. }
}

There are 3 sections in match. Apart from the winners, we also make a draw query in ‘None’. Because; The match may also end in a draw. After the last score is printed, we ask the user a question again. Like replaying or resetting the whole game. If replay is selected, a new game is opened without resetting the score. If Reset is selected, the new game score is unlocked by resetting. The same question is asked in the case of a draw.