tic_tac_toe::board

Struct Board

pub struct Board {
    size: (usize, usize),
    cells: Cells,
}

A two-dimensional game board.

Fields
size: (usize, usize)

Board dimensions: (rows, columns).

cells: Cells

Board cells stored in a 1D vector.

Implementations

impl Board

pub fn get_legal_moves(&self) -> Option<Vec<(usize, usize)>>

Returns a vector of all possible moves.

pub fn is_legal_move(&self, pos: (usize, usize)) -> bool

Checks if the move is legal.

pub fn make_move(&mut self, pos: (usize, usize), val: Cell) -> bool

Tries to make a move.

pub fn check_state(&self, sequence: usize) -> GameState

Checks the actual state of the game.

pub fn test_rules()

Some simple tests.

impl Board

pub fn new(rows: usize, cols: usize) -> Result<Self, Error>

Creates a new game board with the given dimensions.

pub fn clear(&mut self)

Resets all cells to empty.

pub fn get_size(&self) -> (usize, usize)

Returns the board dimensions as (rows, columns).

pub fn get_cells(&self) -> &Vec<Cell>

Returns the board as 1D vector of cell states.

pub fn get_cells_mut(&mut self) -> &mut Vec<Cell>

Returns the board as 1D mut vector of cell states.

pub fn is_empty(&self) -> bool

Returns true if all cells are empty.

pub fn is_full(&self) -> bool

Returns true if all cells are full.

pub fn is_cell_empty(&self, row: usize, col: usize) -> bool

Returns true if the given cell is empty.

pub fn get(&self, row: usize, col: usize) -> Option<&Cell>

Returns the reference of the given cell, or None if out of bounds.

pub fn get_mut(&mut self, row: usize, col: usize) -> Option<&mut Cell>

Returns the mut reference of the given cell, or None if out of bounds.

pub fn set(&mut self, row: usize, col: usize, status: Cell) -> bool

Attempts to set a cell to the given value. Returns true if the operation succeeded.

pub fn row_iter(&self, row: usize) -> impl Iterator<Item = &Cell>

Returns an iterator over the given row.

pub fn all_rows(&self) -> Vec<Vec<&Cell>>

Returns a vector of all rows.

pub fn col_iter(&self, col: usize) -> impl Iterator<Item = &Cell>

Returns an iterator over the given column.

pub fn all_cols(&self) -> Vec<Vec<&Cell>>

Returns a vector of all columns.

pub fn diag_iter( &self, pos: (usize, usize), rev: bool, ) -> impl Iterator<Item = &Cell>

Returns an iterator over the given diagonal. pos - start position (cell) rev - from right to left diagonal

pub fn all_diags(&self) -> Vec<Vec<&Cell>>

Returns a vector of all diagonals.

pub fn print(&self, setup: &PrintSetup)

Pretty-prints the board using the given setup.

pub fn test_board(size: (usize, usize)) -> Result<(), String>

Some simple tests.

fn idx(&self, row: usize, col: usize) -> usize

Converts a (row, col) pair into a 1D index.